reorganize and minor fixes
